Receive mode now ignores duplicate reports and logs communication

+++ serverDataTransfer/communications/commsSafety.c
@@ -100,10 +100,8 @@ long getLastCommunication(int serverId){
 
                 number = atoi(val2);
 
-                if (number == serverId) stamp = atoi(val);
-
+                if (number == serverId) stamp = atol(val);
             }
-
         }
     }
